+++ backend/funco-follow/src/main/java/com/found_404/funco/follow/dto/RatioPrice.java
@@ -2,6 +2,6 @@ package com.found_404.funco.follow.dto;
 
 public record RatioPrice(
         Double ratio,
-        Long price
+        Double price
 ) {
 }
